# docs/design/openapi.yaml
openapi: 3.1.0
info:
  title: Multi-Agent AI Platform API
  version: "1.0.0"
  description: >
    Modular Monolith · Hexagonal · Plugin · Event-Driven.
    Auth: Firebase ID Token (local JWT verify). Tenant (workspace_id) derived from identity, never client-supplied.
    Errors: RFC 9457 Problem Details. Pagination: cursor-based. JSON: snake_case.
    All collection endpoints are wrapped as { data: [...], meta: { next_cursor, limit } } (API-04);
    bounded lists set next_cursor: null. Every response returns X-Correlation-Id.
    The interactive WebSocket endpoint /api/v1/ws is out of scope for OpenAPI and is specified in 03-api-spec.md §3.2.
servers:
  - url: /api/v1
security:
  - firebaseBearer: []
...
paths:
...
  /knowledge/search:
    post:
      tags: [knowledge]
      operationId: searchKnowledge
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/KnowledgeSearchIn" } } } }
      responses:
        "200": { description: OK, headers: { X-Correlation-Id: { $ref: "#/components/headers/CorrelationId" } }, content: { application/json: { schema: { $ref: "#/components/schemas/RetrievedChunkPage" } } } }
        default: { $ref: "#/components/responses/Problem" }
  /knowledge/documents:
    get:
      tags: [knowledge]
      operationId: listDocuments
      parameters: [ { $ref: "#/components/parameters/SpaceId" }, { $ref: "#/components/parameters/Limit" }, { $ref: "#/components/parameters/Cursor" } ]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/DocumentPage" } } } }
        default: { $ref: "#/components/responses/Problem" }
  /knowledge/documents/{id}:
    get:
      tags: [knowledge]
      operationId: getDocument
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/DocumentOut" } } } }
        default: { $ref: "#/components/responses/Problem" }
  /knowledge/reindex:
    post:
      tags: [knowledge]
      operationId: reindexDocuments
      description: "Rebuild the index for one or more documents. **Re-indexing never resets a document** (INV-K3): each target is REPLACED by a brand-new `pending` document over the same `file_id`, and the old one — its Qdrant points, its chunk rows, its row — is destroyed first, in that order. Destroying it up front is what keeps the corpus free of duplicates: point ids derive from the document id (`chunk_point_id`), so a second document over the same file would answer every search twice, forever. The cost is stated plainly: **the file is not searchable until the job finishes**, and a job that is cancelled or fails leaves it that way until it is re-indexed again. Only TERMINAL documents (`indexed`/`failed`) may be re-indexed — a `pending`/`indexing` one is 409, both because the rebuild it is asking for is already under way and because deleting a row a worker is mid-pipeline on would break its chunk writes. Unknown or another tenant's id ⇒ 404; the ids are de-duplicated, and 1..50 of them are accepted (a bigger corpus is re-indexed in batches, so one request can never delete an unbounded number of vectors). Answers **202** with the job: the work is a worker's, and 201 would promise a finished rebuild. `Idempotency-Key` is honoured and matters here more than anywhere — a duplicated POST would destroy and rebuild a second time, at the workspace's expense. Permission `knowledge:manage`, not `knowledge:read`: this deletes a working index and spends embedding quota."
      parameters: [ { $ref: "#/components/parameters/IdempotencyKey" } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ReindexIn" } } } }
      responses:
        "202": { description: Accepted, content: { application/json: { schema: { $ref: "#/components/schemas/ReindexJobOut" } } } }
        default: { $ref: "#/components/responses/Problem" }
  /knowledge/reindex/{id}:
    get:
      tags: [knowledge]
      operationId: getReindexJob
      description: "One re-index job's progress. **Nothing about the progress is stored**: the job row carries only its identity and its `cancelled_at`, and every number here is DERIVED by reading the status of the documents the job created. A stored counter would need a writer on the worker's side and would drift the moment one crashed; a derived one cannot disagree with the corpus it describes. `items` is the per-file phase list (one entry per document, bounded by the request's own 50), `percent` counts terminal items, and `current_file_id` names the file whose document is `indexing` right now — a reference, not a name, exactly like the pin list: the client already joins `GET /files` by id. Permission `knowledge:read`."
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/ReindexJobOut" } } } }
        default: { $ref: "#/components/responses/Problem" }
  /knowledge/reindex/{id}/cancel:
    post:
      tags: [knowledge]
      operationId: cancelReindexJob
      description: "Stop a running re-index. **Cancellation is a claim on the documents the worker has not started yet**, and it works through the lifecycle that already exists rather than a new one: every `pending` document of this job is moved to `failed` with the reason recorded, which is precisely the terminal state the worker's own DD-09 redelivery guard declines to run against — so the queued event arrives and does nothing, with no worker change and no way to un-publish a message. A document already `indexing` is NOT stopped: its pipeline is mid-flight in another process, it keeps its own outcome, and the response shows it still running so the client is never told a lie. Cancelling twice is a no-op that answers 200 without writing; cancelling a job whose items are all terminal is 409 — the work it asks to stop has already happened. The destroyed original index is NOT restored: cancelling leaves those files unindexed, which is the same cost the POST already stated. Permission `knowledge:manage`."
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      responses:
        "200": { description: Cancelled, content: { application/json: { schema: { $ref: "#/components/schemas/ReindexJobOut" } } } }
        default: { $ref: "#/components/responses/Problem" }
...
  /knowledge/documents/{id}/summary/export:
    get:
      tags: [knowledge]
      operationId: exportDocumentSummary
      description: "Download the stored summary as a PDF or DOCX (BE-RAG-012). **A GET returning bytes, not a job — a deliberate departure from the plan's own proposal.** BE-RAG-012 asked for a pollable, downloadable export job with an optional synchronous path for small summaries; every summary IS small, because the pipeline caps its reduce step, so the largest body this can be handed is a few tens of kilobytes of Markdown. The job machinery — a row, an event, a worker, an object in MinIO, a presigned URL, a polling loop and an expiry policy — would all exist to manage work that finishes before the client's first poll. Alpha needed it because Alpha's export ran over unbounded text; ours is bounded by contract, so the optional path is the only path. GET is then the honest verb: this is another representation of a resource that already exists, it changes nothing, and it can be retried freely. The render runs in a worker thread, so one export never becomes every other request's latency. `filename` is the source file's own name, sent by the client so the saved document is recognisable and so the exported page can be headed with it; it is REJECTED rather than escaped when it carries a path separator, a control character or more than 120 characters, because a name that has to be rewritten to be safe is a name the client should not have sent. The header uses RFC 6266's `filename*=UTF-8''` form: these names are Arabic more often than not, and the plain form is latin-1 only. Right-to-left layout follows the summary's `lang`, except for `auto`, where the majority script of the text decides — an English summary quoting one Arabic term is an English document. No stored summary under the key ⇒ 404, from the same read the JSON `GET` uses. Permission `knowledge:read`, not `knowledge:manage`: this produces no new artefact and spends no model budget."
      parameters:
        - { $ref: "#/components/parameters/Id" }
        - { name: kind, in: query, required: true, schema: { type: string, enum: [overview, full] } }
        - { name: lang, in: query, required: true, schema: { type: string, enum: [auto, ar, en] } }
        - { name: format, in: query, required: true, schema: { type: string, enum: [pdf, docx] } }
        - { name: filename, in: query, required: false, schema: { type: string, default: summary, maxLength: 120 } }
      responses:
        "200":
          description: The rendered document
          headers:
            Content-Disposition:
              description: "`attachment; filename*=UTF-8''<name>.<format>`"
              schema: { type: string }
          content:
            application/pdf: { schema: { type: string, format: binary } }
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema: { type: string, format: binary }
        default: { $ref: "#/components/responses/Problem" }
  /knowledge/summary-jobs/{id}:
    get:
      tags: [knowledge]
      operationId: getSummaryJob
      description: "One summary build's progress (BE-RAG-009/011). **Unlike a re-index job, every number here is READ FROM THE ROW rather than derived**, and the divergence is deliberate: a re-index derives its progress because the corpus already records it — its items point at documents that carry their own statuses — while a summary has no second witness until it exists, and \"map step 7 of 42\" is known only to the worker doing step 7. There is nothing to derive it from, so the row counts, and the honest cost is stated where the counter lives: a worker that dies mid-run leaves `done_chunks` stale until redelivery restarts the build. `percent` is the aggregate's own, not the client's arithmetic — a succeeded job is 100 even though the reduce step it just finished was never counted in `total_chunks`. `error` IS published, where `DocumentOut` withholds indexing's: a summary failure is a sentence this module wrote for this reader (\"no indexed text\", \"cancelled\"), not a provider's raw output. Unknown or another tenant's ⇒ 404. Permission `knowledge:read` — watching a build is reading; only starting and stopping one is privileged."
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/SummaryJobOut" } } } }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    firebaseBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
    Limit: { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    Cursor: { name: cursor, in: query, required: false, schema: { type: string } }
...
    SpaceId: { name: space_id, in: query, required: true, schema: { type: string, format: uuid } }
...
  headers:
    CorrelationId:
      description: "Request correlation id, echoed on every response (generated if absent). See 03 §0."
      schema: { type: string }
  responses:
    Problem:
      description: RFC 9457 Problem Details
      headers:
        X-Correlation-Id: { $ref: "#/components/headers/CorrelationId" }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    KnowledgeSearchIn:
      type: object
      required: [query]
      properties: { query: {type: string}, k: {type: integer, maximum: 50, default: 5} }
...
    DocumentPage:
      type: object
      properties: { data: { type: array, items: { $ref: "#/components/schemas/DocumentOut" } }, meta: { $ref: "#/components/schemas/PageMeta" } }
...
    RetrievedChunkPage:
      type: object
      properties: { data: { type: array, items: { $ref: "#/components/schemas/RetrievedChunkOut" } }, meta: { $ref: "#/components/schemas/PageMeta" } }
...